Add lint tools to the project
Added two Makefile recipes in order to format the code and pass the
linter and the type checker using autopep8 for formatting and mypy and
flake for linting.

This will ensure that all contributions to this project will follow the
same rules and the code will be more traceble and easier to read and
follow.

As part of this PR there are some other minor fixes:

* Removed `*` imports
* Fixed some naming conventions
* Fixed some Imports between SqlModel/SqlAlchemy objects
* Added type returns to all the functions
* Removed unused imports
* Fix minor style issues

Signed-off-by: Oscar Moya <oscar.moya@alefedge.com>